+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 ## CHANGELOG
 
-1. some refactoring.
-2. improved subreddit names debounced search.
+1. added skyblue `OP` text next to comments of OP.